+++ Dyalect/Compiler/BuilderOptions.cs
@@ -6,7 +6,9 @@
         {
             Debug = false,
             NoLangModule = false,
-            NoWarnings = false
+            NoWarnings = false,
+            NoWarningsLinker = false,
+            LinkerSkipChecksum = false
         };
 
         public bool Debug { get; set; }
@@ -14,5 +16,9 @@
         public bool NoLangModule { get; set; }
 
         public bool NoWarnings { get; set; }
+
+        public bool NoWarningsLinker { get; set; }
+
+        public bool LinkerSkipChecksum { get; set; }
     }
 }