+++ Dyalect/Compiler/Unit.cs
@@ -46,6 +46,8 @@ namespace Dyalect.Compiler
             TypeMap = unit.TypeMap;
         }
 
+        internal int Checksum { get; set; }
+
         internal Unit Clone(DebugInfo di) => new Unit(this, di);
 
         internal int Id { get; set; }