+++ Dyalect/Linker/LinkerWarning.cs
@@ -1,0 +1,7 @@
+﻿namespace Dyalect.Linker
+{
+    public enum LinkerWarning
+    {
+        NewerSourceFile = 500
+    }
+}