+++ Dyalect/Strings/LinkerErrors.Designer.cs
@@ -79,6 +79,15 @@ namespace Dyalect.Strings {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Checksum validation of dependency &quot;{0}&quot; failed for module &quot;{1}&quot;. .
+        /// </summary>
+        internal static string ChecksumValidationFailed {
+            get {
+                return ResourceManager.GetString("ChecksumValidationFailed", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Duplicate name of a module &quot;{1}&quot; in an assembly &quot;{0}&quot;..
         /// </summary>
         internal static string DuplicateModuleName {
@@ -102,6 +111,15 @@ namespace Dyalect.Strings {
         internal static string ModuleNotFound {
             get {
                 return ResourceManager.GetString("ModuleNotFound", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to A newer version of source file is available for module &quot;{0}&quot;..
+        /// </summary>
+        internal static string NewerSourceFile {
+            get {
+                return ResourceManager.GetString("NewerSourceFile", resourceCulture);
             }
         }