+++ Docs/Changes.md
@@ -1,3 +1,7 @@
+# 0.9.14
+  * Change: Multiple runtime error messages are corrected.
+  * Fixed: _Incorrect exception message for custom types for `IndexOutOfRange`_ ([Issue #253](https://github.com/vorov2/dyalect/issues/253)).
+  
 # 0.9.13
   * Done: _Pattern match validation_ ([Issue #241](https://github.com/vorov2/dyalect/issues/241)).
   * Done: _Tuple pattern optimization_ ([Issue #244](https://github.com/vorov2/dyalect/issues/244)).